+++ aggregator/src/aggregation.rs
@@ -1,5 +1,5 @@
 /// Config to constrain batch data (decoded blob data)
-mod batch_data;
+pub mod batch_data;
 /// Circuit implementation of aggregation circuit.
 mod circuit;
 /// Config for aggregation circuit
@@ -11,6 +11,7 @@ mod rlc;
 /// Utility module
 mod util;
 
+pub use batch_data::BatchData;
 pub(crate) use batch_data::BatchDataConfig;
 pub use decoder::decode_bytes;
 pub(crate) use decoder::{witgen, DecoderConfig, DecoderConfigArgs};
