+++ aggregator/src/data_availability/eip4844/blob_data.rs
@@ -10,8 +10,11 @@ use itertools::Itertools;
 use zkevm_circuits::{table::U8Table, util::Challenges};
 
 use crate::{
-    aggregation::POWS_OF_256,
-    data_availability::eip4844::blob::{BLOB_WIDTH, N_BLOB_BYTES, N_DATA_BYTES_PER_COEFFICIENT},
+    aggregation::{
+        batch_data::{N_BLOB_BYTES, N_DATA_BYTES_PER_COEFFICIENT},
+        POWS_OF_256,
+    },
+    data_availability::eip4844::blob::BLOB_WIDTH,
     RlcConfig,
 };
